requirements.txt
flask==2.3.3
openai==0.28.0
python-dotenv==1.0.0
flask-session==0.5.0
